feat: 쿠폰 발급 api 추가 (#44)
* chore: api 모듈 rename (api -> api-fcfs)

선착순 처리와 쿠폰 처리 application 분리 위함

* chore: api-coupon 모듈 추가

coupon의 발급과 발송을 담당하는 web application module
