+++ fcfs-coupon-api-coupon/src/main/java/com/coupop/fcfscoupon/CouponApplication.java
@@ -1,0 +1,12 @@
+package com.coupop.fcfscoupon;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CouponApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CouponApplication.class, args);
+    }
+}
