+++ fcfs-coupon-api-coupon/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.profiles.include=coupon
+
+server.port=8081
+server.shutdown=graceful
+spring.lifecycle.timeout-per-shutdown-phase=20s
