+++ fcfs-coupon-api-coupon/src/test/java/com/coupop/fcfscoupon/CouponApplicationTests.java
@@ -1,0 +1,13 @@
+package com.coupop.fcfscoupon;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = CouponApplication.class)
+class CouponApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
