+++ fcfs-coupon-api-coupon/src/test/java/com/coupop/fcfscoupon/api/coupon/CouponControllerTest.java
@@ -1,0 +1,70 @@
+package com.coupop.fcfscoupon.api.coupon;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import com.coupop.fcfscoupon.api.coupon.dto.IssuanceRequest;
+import com.coupop.fcfscoupon.domain.coupon.CouponService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.ResultActions;
+
+@WebMvcTest
+class CouponControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ObjectMapper objectMapper;
+
+    @MockBean
+    private CouponService couponService;
+
+    @DisplayName("쿠폰 발행에 성공하면 Accepted 상태를 반환한다.")
+    @Test
+    void issue() throws Exception {
+        // given
+        final IssuanceRequest request = new IssuanceRequest(1L, "foo@bar.com");
+
+        // when
+        final ResultActions resultActions = performPost("/issue", request);
+
+        // then
+        resultActions
+                .andExpect(status().isAccepted());
+    }
+
+    @DisplayName("쿠폰 발행시, 이메일이 형식에 맞지 않으면 Bad Request 상태를 반환한다.")
+    @ParameterizedTest
+    @ValueSource(strings = {"foobar.com", "foo@", "foo@com"})
+    void issue_responseError_ifEmailInvalid(final String invalidEmail) throws Exception {
+        // given
+        final IssuanceRequest request = new IssuanceRequest(1L, invalidEmail);
+
+        // when
+        final ResultActions resultActions = performPost("/issue", request);
+
+        // then
+        resultActions
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("title").value("형식에 맞는 이메일을 입력하세요."));
+    }
+
+    private ResultActions performPost(final String url, final Object request) throws Exception {
+        return mockMvc.perform(post(url)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(request)))
+                .andDo(print());
+    }
+}
