+++ fcfs-coupon-api-fcfs/src/main/java/com/coupop/fcfscoupon/api/fcfs/dto/HistoryResponse.java
@@ -1,0 +1,14 @@
+package com.coupop.fcfscoupon.api.fcfs.dto;
+
+import com.coupop.fcfscoupon.domain.coupon.model.CouponIssueHistory;
+import java.util.List;
+
+public record HistoryResponse(List<IssuedCouponResponse> issuedCoupons) {
+
+    public static HistoryResponse of(final List<CouponIssueHistory> history) {
+        List<IssuedCouponResponse> issuedCoupons = history.stream()
+                .map(IssuedCouponResponse::of)
+                .toList();
+        return new HistoryResponse(issuedCoupons);
+    }
+}
