+++ fcfs-coupon-api-fcfs/src/main/java/com/coupop/fcfscoupon/api/fcfs/dto/IssuedCouponResponse.java
@@ -1,0 +1,10 @@
+package com.coupop.fcfscoupon.api.fcfs.dto;
+
+import com.coupop.fcfscoupon.domain.coupon.model.CouponIssueHistory;
+
+public record IssuedCouponResponse(String id, String date) {
+
+    public static IssuedCouponResponse of(final CouponIssueHistory history) {
+        return new IssuedCouponResponse(history.getId(), history.getCreatedAt());
+    }
+}
