+++ fcfs-coupon-api-fcfs/src/main/java/com/coupop/fcfscoupon/api/fcfs/dto/ResendRequest.java
@@ -1,0 +1,4 @@
+package com.coupop.fcfscoupon.api.fcfs.dto;
+
+public record ResendRequest(String historyId) {
+}
