+++ fcfs-coupon-api-fcfs/src/test/java/com/coupop/fcfscoupon/FcfsCouponApplicationTests.java
@@ -1,0 +1,13 @@
+package com.coupop.fcfscoupon;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = FcfsCouponApplication.class)
+class FcfsCouponApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
