+++ fcfs-coupon-api-fcfs/src/test/java/com/coupop/fcfscoupon/api/fcfs/testconfig/MongoDatabaseCleaner.java
@@ -1,0 +1,20 @@
+package com.coupop.fcfscoupon.api.fcfs.testconfig;
+
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MongoDatabaseCleaner {
+
+    private final MongoTemplate mongoTemplate;
+
+    public MongoDatabaseCleaner(final MongoTemplate mongoTemplate) {
+        this.mongoTemplate = mongoTemplate;
+    }
+
+    public void clean() {
+        for (String collectionName : mongoTemplate.getCollectionNames()) {
+            mongoTemplate.dropCollection(collectionName);
+        }
+    }
+}
