--- fcfs-coupon-api/src/main/java/com/coupop/fcfscoupon/api/dto/HistoryRequest.java
@@ -1,7 +1,0 @@
-package com.coupop.fcfscoupon.api.dto;
-
-import com.coupop.fcfscoupon.common.inputvalidation.EmailValidation;
-import jakarta.validation.constraints.Email;
-
-public record HistoryRequest(@Email(regexp = EmailValidation.REGEX, message = EmailValidation.MESSAGE) String email) {
-}
