--- fcfs-coupon-api/src/main/java/com/coupop/fcfscoupon/api/dto/IssuanceRequest.java
@@ -1,7 +1,0 @@
-package com.coupop.fcfscoupon.api.dto;
-
-import com.coupop.fcfscoupon.common.inputvalidation.EmailValidation;
-import jakarta.validation.constraints.Email;
-
-public record IssuanceRequest(@Email(regexp = EmailValidation.REGEX, message = EmailValidation.MESSAGE) String email) {
-}
