--- fcfs-coupon-api/src/main/java/com/coupop/fcfscoupon/api/dto/ResendRequest.java
@@ -1,4 +1,0 @@
-package com.coupop.fcfscoupon.api.dto;
-
-public record ResendRequest(String historyId) {
-}
