+++ settings.gradle.kts
@@ -1,6 +1,7 @@
 rootProject.name = "fcfs-coupon"
-include("fcfs-coupon-api")
-include("fcfs-coupon-mq")
+include("fcfs-coupon-api-fcfs")
+include("fcfs-coupon-api-coupon")
 include("fcfs-coupon-domain-fcfs")
 include("fcfs-coupon-domain-coupon")
 include("fcfs-coupon-common")
+include("fcfs-coupon-mq")
